0.1.4_beta_1

+++ src/main/java/me/kirito5572/listener/filterListener.java
@@ -1,6 +1,6 @@
 package me.kirito5572.listener;
 
-import me.kirito5572.App;
+import me.kirito5572.objects.FilterSystem;
 import me.kirito5572.objects.MySQLConnector;
 import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.Guild;
@@ -11,38 +11,27 @@ import net.dv8tion.jda.api.events.message.guild.GuildMessageReceivedEvent;
 import net.dv8tion.jda.api.events.message.guild.GuildMessageUpdateEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 import org.jetbrains.annotations.NotNull;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
-import java.sql.ResultSet;
 import java.util.Objects;
 import java.util.concurrent.TimeUnit;
 
 public class filterListener extends ListenerAdapter {
-    private static String[] filter_data;
-    private static String[][] white_list_data;
+    private final Logger logger = LoggerFactory.getLogger(filterListener.class);
     private final MySQLConnector mySqlConnector;
+    private final FilterSystem filterSystem;
 
-    public filterListener(MySQLConnector mySqlConnector) {
+    public filterListener(MySQLConnector mySqlConnector, FilterSystem filterSystem) {
         this.mySqlConnector = mySqlConnector;
+        this.filterSystem = filterSystem;
     }
 
     @Override
     public void onReady(@NotNull ReadyEvent event) {
-        if (getFilterDataFromDB(mySqlConnector)) return;
-        try (ResultSet resultSet = mySqlConnector.Select_Query("SELECT * FROM blitz_bot.WhiteListWord;", new int[0], new String[0])) {
-            if (resultSet.last()) {
-                return;
-            }
-            String[][] a = new String[resultSet.getRow()][2];
-            resultSet.first();
-            int i = 0;
-            while (resultSet.next()) {
-                a[i][0] = resultSet.getString("FilterWord");
-                a[i][1] = resultSet.getString("Word");
-            }
-            filterListener.setWhite_list_data(a);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
+        filterSystem.filterRefresh();
+        filterSystem.whiteFilterRefresh();
+        logger.info("필터 ONLINE");
     }
 
     @Override
@@ -60,33 +49,6 @@ public class filterListener extends ListenerAdapter {
     }
 
     /**
-     * Get prohibited word from DataBase
-     *
-     * @param mySqlConnector {@link MySQLConnector}
-     *
-     * @return if true, get data success
-     *         if false, get data fail
-     */
-
-    public static boolean getFilterDataFromDB(@NotNull MySQLConnector mySqlConnector) {
-        try (ResultSet resultSet = mySqlConnector.Select_Query("SELECT * FROM blitz_bot.FilterWord;", new int[0], new String[0])) {
-            if (resultSet.last()) {
-                return true;
-            }
-            String[] a = new String[resultSet.getRow()];
-            resultSet.first();
-            int i = 0;
-            while (resultSet.next()) {
-                a[i] = resultSet.getString("Word");
-            }
-            filterListener.setFilter_data(a);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        return false;
-    }
-
-    /**
      *
      * Thread execution to check prohibited words
      * @param member the member who need check he/her message
@@ -97,7 +59,6 @@ public class filterListener extends ListenerAdapter {
     private void filter_data(@NotNull Member member,@NotNull Message message) {
         Guild guild = member.getGuild();
         if (guild.getId().equals("826704284003205160")) {
-            filter_data = App.getFilterList();
             try {
                 String message_raw = message.getContentRaw();
                 if (message_raw.length() > 1) {
@@ -120,17 +81,6 @@ public class filterListener extends ListenerAdapter {
 
     private void filter(@NotNull Member member, @NotNull Message message) {
         String rawMessage = message.getContentRaw();
-        if (member.getPermissions().contains(Permission.ADMINISTRATOR) |
-                member.getPermissions().contains(Permission.MESSAGE_MANAGE) |
-                member.getPermissions().contains(Permission.KICK_MEMBERS) |
-                member.getPermissions().contains(Permission.BAN_MEMBERS) |
-                member.getPermissions().contains(Permission.MANAGE_PERMISSIONS) |
-                member.getPermissions().contains(Permission.MANAGE_CHANNEL) |
-                member.getPermissions().contains(Permission.MANAGE_EMOTES) |
-                member.getPermissions().contains(Permission.MANAGE_SERVER) |
-                member.getPermissions().contains(Permission.MANAGE_WEBHOOKS)) {
-            return;
-        }
         if (member.getUser().isBot()) {
             return;
         }
@@ -139,9 +89,9 @@ public class filterListener extends ListenerAdapter {
         rawMessage = rawMessage.replaceAll(match, "");
         boolean filter = false;
         boolean filter_continue = false;
-        for (String data : filter_data) {
+        for (String data : filterSystem.getFilterList()) {
             if (rawMessage.contains(data)) {
-                for(String[] a : white_list_data) {
+                for(String[] a : filterSystem.getWhiteFilterList()) {
                     if(data.equals(a[0])) {
                         if(rawMessage.contains(a[1])) {
                             filter_continue = true;
@@ -158,30 +108,20 @@ public class filterListener extends ListenerAdapter {
             }
         }
         if(filter) {
+            if (member.getPermissions().contains(Permission.ADMINISTRATOR) |
+                    member.getPermissions().contains(Permission.MESSAGE_MANAGE) |
+                    member.getPermissions().contains(Permission.KICK_MEMBERS) |
+                    member.getPermissions().contains(Permission.BAN_MEMBERS) |
+                    member.getPermissions().contains(Permission.MANAGE_PERMISSIONS) |
+                    member.getPermissions().contains(Permission.MANAGE_CHANNEL) |
+                    member.getPermissions().contains(Permission.MANAGE_EMOTES) |
+                    member.getPermissions().contains(Permission.MANAGE_SERVER) |
+                    member.getPermissions().contains(Permission.MANAGE_WEBHOOKS)) {
+                logger.info("특정 등급 이상 권한 부여자가 필터링에 걸리는 단어를 사용하였으나 통과되었습니다.");
+                return;
+            }
             message.getTextChannel().deleteMessageById(message.getId()).complete();
             message.getTextChannel().sendMessage(member.getAsMention() + ", 금지어 사용에 주의하여주십시오.").complete().delete().queueAfter(10, TimeUnit.SECONDS);
         }
-    }
-
-    /**
-     * Setting prohibited word list
-     *
-     * @param filter_data prohibited data list
-     *
-     */
-
-    public static void setFilter_data(@NotNull String[] filter_data) {
-        filterListener.filter_data = filter_data;
-    }
-
-    /**
-     * Setting non-prohibited word list
-     *
-     * @param white_list_data non-prohibited data list
-     *
-     */
-
-    public static void setWhite_list_data(@NotNull String[][] white_list_data) {
-        filterListener.white_list_data = white_list_data;
     }
 }
